getmapping 분리 완료
