+++ app/src/main/java/com/codesoom/assignment/task/TaskRepository.java
@@ -1,5 +1,6 @@
 package com.codesoom.assignment.task;
 
+import com.codesoom.assignment.exception.ResourceNotFoundException;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
@@ -8,7 +9,19 @@ import java.util.List;
 @Repository
 public class TaskRepository {
 
-    private List<Task> tasks = new ArrayList<>();
-    private Long id = 1L;
+    public List<Task> tasks = new ArrayList<>();
+    public Long id = 1L;
+
+    public Long nextId(){
+        return id++;
+    }
+
+    public Task findTaskById(Long id) throws ResourceNotFoundException {
+        for (Task task : tasks){
+            if (task.getId() == id)
+                return task;
+        }
+        throw new ResourceNotFoundException("no task id : " + id);
+    }
 
 }
