+++ app/src/main/java/com/codesoom/assignment/task/TaskService.java
@@ -2,6 +2,8 @@ package com.codesoom.assignment.task;
 
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 public class TaskService {
 
@@ -9,5 +11,13 @@ public class TaskService {
 
     public TaskService(TaskRepository taskRepository) {
         this.taskRepository = taskRepository;
+    }
+
+    public List<Task> getTasks() {
+        return taskRepository.tasks;
+    }
+
+    public Task getTaskById(Long id) {
+        return taskRepository.findTaskById(id);
     }
 }
